some cleanup, and moving testing scripts to already existing coinbase script
